+++ @navikt/core/react/src/form/combobox/FilteredOptions/FilteredOptions.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from "react";
+import React, { useMemo, useRef, useState } from "react";
 import { useClientLayoutEffect } from "../../../utils-external";
 import { DismissableLayer } from "../../../utils/components/dismissablelayer/DismissableLayer";
 import { Floating } from "../../../utils/components/floating/Floating";
@@ -29,6 +29,7 @@ const FilteredOptions = () => {
     isValueNew,
     toggleIsListOpen,
   } = useFilteredOptionsContext();
+
   const [localOpen, setLocalOpen] = useState(isListOpen);
 
   const floatingRef = useRef<HTMLDivElement | null>(null);
@@ -42,16 +43,30 @@ const FilteredOptions = () => {
     queueMicrotask(() => setLocalOpen(isListOpen));
   }, [isListOpen]);
 
-  const { maxSelected, isMultiSelect } = useSelectedOptionsContext();
+  const { maxSelected, isMultiSelect, selectedOptions } =
+    useSelectedOptionsContext();
 
   const shouldRenderNonSelectables =
     (isMultiSelect && maxSelected.limit) || // Render maxSelected message
     isLoading || // Render loading message
     (!isLoading && filteredOptions.length === 0 && !allowNewValues); // Render no hits message
 
+  const singleSelectedOption = useMemo(() => {
+    if (
+      isMultiSelect ||
+      selectedOptions.length === 0 ||
+      selectedOptions.length > 1
+    ) {
+      return null;
+    }
+
+    return selectedOptions[0];
+  }, [isMultiSelect, selectedOptions]);
+
   const shouldRenderFilteredOptionsList =
     (allowNewValues && isValueNew && !maxSelected.isLimitReached) || // Render add new option
-    filteredOptions.length > 0; // Render filtered options
+    filteredOptions.length > 0 || // Render filtered options
+    !!singleSelectedOption;
 
   return (
     <DismissableLayer
@@ -104,14 +119,26 @@ const FilteredOptions = () => {
             {isValueNew && !maxSelected.isLimitReached && allowNewValues && (
               <AddNewOption />
             )}
-            {filteredOptions.map((option) => (
-              <FilteredOptionsItem key={option.value} option={option} />
-            ))}
+            {singleSelectedOption && (
+              <FilteredOptionsItem
+                key={singleSelectedOption.value}
+                option={singleSelectedOption}
+              />
+            )}
+            {filteredOptions.map((option) => {
+              if (
+                singleSelectedOption &&
+                option.value === singleSelectedOption.value
+              ) {
+                return null;
+              }
+
+              return <FilteredOptionsItem key={option.value} option={option} />;
+            })}
           </ul>
         )}
       </Floating.Content>
     </DismissableLayer>
   );
 };
-
 export default FilteredOptions;